# src/swagger/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"UserResponse": {
				"properties": {
					"id": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"nickname": {
						"type": "string"
					},
					"provider": {
						"type": "string"
					},
					"profileImage": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"email",
					"provider"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_UserResponse_": {
				"description": "통일된 API 응답 형식",
				"properties": {
					"isSuccess": {
						"type": "boolean"
					},
					"code": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"result": {
						"allOf": [
							{
								"$ref": "#/components/schemas/UserResponse"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"isSuccess",
					"code",
					"message",
					"result"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_UserResponse-Array_": {
				"description": "통일된 API 응답 형식",
				"properties": {
					"isSuccess": {
						"type": "boolean"
					},
					"code": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"result": {
						"items": {
							"$ref": "#/components/schemas/UserResponse"
						},
						"type": "array",
						"nullable": true
					}
				},
				"required": [
					"isSuccess",
					"code",
					"message",
					"result"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TripScheduleResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"destination": {
						"type": "string"
					},
					"destinationType": {
						"type": "string"
					},
					"startDate": {
						"type": "string"
					},
					"endDate": {
						"type": "string"
					},
					"createdBy": {
						"type": "string"
					},
					"members": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"title",
					"destination",
					"destinationType",
					"startDate",
					"endDate",
					"createdBy",
					"members"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ApiResponse_TripScheduleResponse-Array_": {
				"description": "통일된 API 응답 형식",
				"properties": {
					"isSuccess": {
						"type": "boolean"
					},
					"code": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"result": {
						"items": {
							"$ref": "#/components/schemas/TripScheduleResponse"
						},
						"type": "array",
						"nullable": true
					}
				},
				"required": [
					"isSuccess",
					"code",
					"message",
					"result"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "access_token",
				"in": "cookie"
			}
		}
	},
	"info": {
		"title": "server",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/users/{id}": {
			"get": {
				"operationId": "GetUserById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_UserResponse_"
								}
							}
						}
					}
				},
				"description": "ID로 유저 조회",
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "유저 삭제",
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/users/email/{email}": {
			"get": {
				"operationId": "FindUserByEmail",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_UserResponse_"
								}
							}
						}
					}
				},
				"description": "이메일로 유저 조회",
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "email",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/users/emails": {
			"post": {
				"operationId": "FindUsersByEmails",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_UserResponse-Array_"
								}
							}
						}
					}
				},
				"description": "이메일 배열로 유저 정보 조회",
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"emails": {
										"items": {
											"type": "string"
										},
										"type": "array"
									}
								},
								"required": [
									"emails"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/users/search": {
			"get": {
				"operationId": "GetUserSearch",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_UserResponse-Array_"
								}
							}
						}
					}
				},
				"description": "이메일로 유저 검색",
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "query",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/users/logout": {
			"post": {
				"operationId": "Logout",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "로그아웃",
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
...
		"/trips/user/{userId}": {
			"get": {
				"operationId": "GetTripsByUserId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_TripScheduleResponse-Array_"
								}
							}
						}
					}
				},
				"description": "유저가 속한 여행 일정 목록 조회",
				"tags": [
					"TripSchedule"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}